+++ test/nodb/src/test/java/org/hibernate/tool/hbm2x/Hbm2EJBDaoTest/TestCase.java
@@ -14,7 +14,7 @@ import javax.persistence.Persistence;
 
 import org.apache.commons.logging.Log;
 import org.hibernate.tool.api.metadata.MetadataDescriptor;
-import org.hibernate.tool.hbm2x.DAOExporter;
+import org.hibernate.tool.internal.export.dao.DAOExporter;
 import org.hibernate.tool.internal.export.pojo.POJOExporter;
 import org.hibernate.tools.test.util.FileUtil;
 import org.hibernate.tools.test.util.HibernateUtil;
